fix: remove check_protected and is_protected
