+++ src/core/jit/model.rs
@@ -287,7 +287,6 @@ pub struct OperationPlan<Input> {
     pub is_introspection_query: bool,
     pub is_dedupe: bool,
     pub is_const: bool,
-    pub is_protected: bool,
     pub min_cache_ttl: Option<NonZeroU64>,
     pub selection: Vec<Field<Input>>,
     pub auth_n: Option<Auth>,
@@ -321,7 +320,6 @@ impl<Input> OperationPlan<Input> {
             is_dedupe: self.is_dedupe,
             is_const: self.is_const,
             min_cache_ttl: self.min_cache_ttl,
-            is_protected: self.is_protected,
             auth_n: self.auth_n,
         })
     }
@@ -348,7 +346,6 @@ impl<Input> OperationPlan<Input> {
             is_dedupe: false,
             is_const: false,
             min_cache_ttl: None,
-            is_protected: false,
             auth_n: None,
         }
     }
