+++ src/core/jit/transform/input_resolver.rs
@@ -75,7 +75,6 @@ where
             is_introspection_query: self.plan.is_introspection_query,
             is_dedupe: self.plan.is_dedupe,
             is_const: self.plan.is_const,
-            is_protected: self.plan.is_protected,
             min_cache_ttl: self.plan.min_cache_ttl,
             selection,
             auth_n: self.plan.auth_n,
