+++ src/core/jit/transform/mod.rs
@@ -2,7 +2,6 @@ mod auth_planer;
 mod check_cache;
 mod check_const;
 mod check_dedupe;
-mod check_protected;
 mod input_resolver;
 mod skip;
 
@@ -10,6 +9,5 @@ pub use auth_planer::*;
 pub use check_cache::*;
 pub use check_const::*;
 pub use check_dedupe::*;
-pub use check_protected::*;
 pub use input_resolver::*;
 pub use skip::*;
